+++ controllers/cluster_controller.go
@@ -64,21 +64,40 @@ func (r *ClusterReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ct
 }
 
 func (r *ClusterReconciler) reconcile(ctx context.Context, cluster *capiv1alpha3.Cluster, log logr.Logger) (ctrl.Result, error) {
+	// ignore cluster from being deleted if ignore annotation is set
 	if _, ok := cluster.Annotations[ignoreClusterDeletion]; ok {
 		IgnoredTotal.WithLabelValues(cluster.Name, cluster.Namespace).Inc()
 		log.Info(fmt.Sprintf("Found annotation %s. Cluster %s/%s will be ignored for deletion", ignoreClusterDeletion, cluster.Namespace, cluster.Name))
 		return ctrl.Result{}, nil
-
 	}
 
+	// check if cluster has a keep-alive label with a valid ISO date string
+	var keepValidExpired bool
+	if v, ok := cluster.Labels[keepValid]; ok {
+		t, err := time.Parse(keepValidTimeLayout, v)
+		if err != nil {
+			ErrorsTotal.WithLabelValues(cluster.Name, cluster.Namespace).Inc()
+			log.Error(err, fmt.Sprintf("failed to parse keep-valid label value for cluster %s/%s", cluster.Namespace, cluster.Name))
+			return ctrl.Result{}, nil
+		}
+		if time.Now().After(t) {
+			keepValidExpired = true
+		} else {
+			IgnoredTotal.WithLabelValues(cluster.Name, cluster.Namespace).Inc()
+			log.Info(fmt.Sprintf("Found label %s. Cluster %s/%s will be ignored for deletion", keepValid, cluster.Namespace, cluster.Name))
+			return ctrl.Result{RequeueAfter: 24 * time.Hour}, nil
+		}
+	}
+
+	// ignore cluster deletion if timestamp is not nil or zero
 	if !cluster.DeletionTimestamp.IsZero() {
 		PendingTotal.WithLabelValues(cluster.Name, cluster.Namespace).Inc()
 		log.Info(fmt.Sprintf("Deletion for cluster %s/%s is already applied", cluster.Namespace, cluster.Name))
 		return ctrl.Result{}, nil
 	}
 
-	// immediately delete the cluster if defaultTTL has passed
-	if deletionTimeReached(cluster) {
+	// immediately delete the cluster if defaultTTL has passed or keep-alive value has expired
+	if deletionTimeReached(cluster) || keepValidExpired {
 		if !r.DryRun {
 			err := r.Client.Delete(ctx, cluster, client.PropagationPolicy(metav1.DeletePropagationBackground))
 			if err != nil {