+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Consider cluster deletion when `keep-valid` label is set.
+
 ## [0.1.0] - 2021-12-07
 
 ### Added